FA-14 - feat(UserProfile): (WIP) Improve and clean the code
