+++ packages/ui_library/lib/widgets/widgets_export.dart
@@ -1,4 +1,3 @@
 export 'u_icon/u_icon_export.dart';
 export 'u_text/u_text_export.dart';
-export 'u_user_profile/u_user_profile_picture.dart';
-export 'u_user_profile/status_indicator.dart';
+export 'u_user_profile/u_user_profile_status.dart';
