+++ packages/ui_showroom/lib/ui_pages/src/u_user_profile_image_page.dart
@@ -116,21 +116,15 @@ class _UserProfileRow extends StatelessWidget {
             const SizedBox.square(
               dimension: 4,
             ),
-            Container(
-                decoration: BoxDecoration(color: Colors.grey.shade300),
-                child: userProfileStatus[0]),
+            userProfileStatus[0],
             const SizedBox.square(
               dimension: 4,
             ),
-            Container(
-                decoration: BoxDecoration(color: Colors.grey.shade300),
-                child: userProfileStatus[1]),
+            userProfileStatus[1],
             const SizedBox.square(
               dimension: 4,
             ),
-            Container(
-                decoration: BoxDecoration(color: Colors.grey.shade300),
-                child: userProfileStatus[2]),
+            userProfileStatus[2],
           ],
         )
       ],
